+++ pkgs/ffigen/test/unit_tests/api_availability_test.dart
@@ -4,7 +4,6 @@
 
 import 'package:ffigen/src/config_provider/config_types.dart';
 import 'package:ffigen/src/header_parser/sub_parsers/api_availability.dart';
-import 'package:pub_semver/pub_semver.dart';
 import 'package:test/test.dart';
 
 void main() {